Move solutions to buried package: project_01
